backend/NewsService/src/test/java/com/news/newsapi/service/test/ApiServiceImplTest.java
//package com.news.newsapi.service.test;
//
//import com.news.newsapi.domain.News;
//import com.news.newsapi.service.ApiService;
//import com.news.newsapi.service.ApiServiceImpl;
//import org.junit.jupiter.api.BeforeEach;
//import org.junit.jupiter.api.Test;
//import org.mockito.InjectMocks;
//import org.mockito.Mock;
//import org.mockito.MockitoAnnotations;
//import org.springframework.http.HttpStatus;
//import org.springframework.http.ResponseEntity;
//import org.springframework.web.client.RestTemplate;
//
//import static org.junit.jupiter.api.Assertions.assertEquals;
//import static org.mockito.ArgumentMatchers.anyString;
//import static org.mockito.ArgumentMatchers.eq;
//import static org.mockito.Mockito.when;
//
//public class ApiServiceImplTest {
//
//    @Mock
//    private RestTemplate restTemplateMock;
//
//    @InjectMocks
//    private ApiService apiService = new ApiServiceImpl();
//
//    @BeforeEach
//    public void setUp() {
//        MockitoAnnotations.initMocks(this);
//    }
//
//    @Test
//    public void testGetNewsApi_Success() {
//        // Arrange
//        News mockedNews = new News(/* Mocked News object */);
//        when(restTemplateMock.getForObject(anyString(), eq(News.class))).thenReturn(mockedNews);
//
//        // Act
//        News result = apiService.getNewsApi();
//
//        // Assert
//        assertEquals(mockedNews, result);
//    }
//
//    @Test
//    public void testGetNewsSearchApi_Success() {
//        // Arrange
//        String keyword = "example";
//        News mockedNews = new News(/* Mocked News object */);
//        when(restTemplateMock.getForObject(anyString(), eq(News.class))).thenReturn(mockedNews);
//
//        // Act
//        News result = apiService.getNewsSearchApi(keyword);
//
//        // Assert
//        assertEquals(mockedNews, result);
//    }
//
//    @Test
//    public void testGetNewsApi_RestTemplateError() {
//        // Arrange
//        when(restTemplateMock.getForObject(anyString(), eq(News.class))).thenReturn(null);
//
//        // Act
//        News result = apiService.getNewsApi();
//
//        // Assert
//        assertEquals(null, result);
//    }
//
//    // Similar test methods can be added for different scenarios
//
//    // Note: These test cases assume a basic structure of the code; additional checks may be needed based on actual implementation.
//}
